requirements.txt
aiogram
fastapi
sqlalchemy
psycopg2-binary
APScheduler
python-dotenv
requests
uvicorn 